fix: automa element selector not working
