+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "automa",
-  "version": "1.11.1",
+  "version": "1.11.2",
   "description": "An extension for automating your browser by connecting blocks",
   "license": "MIT",
   "repository": {
